Added a hash map to cache hash values (<i>v</i> base pairs in length) to seedup <i>u/v</i> heuristic.
<p>The <i>u/v</i> heuristic \b used to generate hash values
(in the \c runHeuristic method) by iterating over the base
pairs in a given EST sequence.  However, this approach turned
out to be rather slow. Furthermore, it was observed that the
same set of hash values were recomputed for various pair-wise
EST comparisons.</p>
        
<p>Therefore, to improve the overall performance of the
<i>u/v</i> heuristic, it was proposed that the hash values be
cached to improve performance. Of course, this does increase
the net amount of memory consumed. Consequently, to aid in
caching only the required subset of EST hash values, this hash
map was introduced to store the necessary sequences and
rapidly access them when needed.</p>

<p>The entries in this hash map are computed by the \c
computeHash method, which is invoked from the runHeuristic
method.</p>