Yet another batch of changes to streamline a lot of the GUI operations. The primary change is streamlining menus and how menus are handled. Still the File menu and data set entry menus need to be handled correctly.
In addition added more checks to ensure that the build environment is 
sane before launching a build. Hopefully this will catch more problems
in the build environment.

Added feature to list jobs/processes running on a machine. This is still
not implemented for windows.

Added feature to display additional information (such as: outputs, errors, 
scripts) for a given jobs