A few more niceities added to Server Connection Tester. Now it displays the default enviorment of the remote server on connection. It will help troubleshoot environmental issues (hopefully).
The server jobs view now shows a suitable message (rather than a blank screen)
when the user does not have any jobs running on the server.

The installer and job details are no longer pushed to stderr but only onto
stdout to resolve hangs on Macs. Corresponding changes to various script files.

Added check to prevent the use of spaces in install path. This was causing
problems in various scripts. Maybe later on we can relax this requirement
and patch up the code path to permit spaces.