Fixed all Javadoc documentation generation issues and added additional files to ensure that Javadoc generates complete documentation for PEACE GUI.
Most of the changes were removing references to the @note tag which is not
supported in Javadoc.

The documentation is currently generated with zero warnings. The configuration
used to generate Javadoc is checked-in in a newly added docs folder as
Javadoc.xml. This XML file is an ant script. However, it may be easier to
generate Javadoc via Eclipse.