Fixed issue where the parameters passed-in to d2 (in the form of command line arguments) were not being parsed correctly.
The source of the issue was the command line parser assumes that the first
argument is the name of the executable, in compliance with the standards.

However, the example(s) were not setting the first parameter to be the
executable name, causing the parameters to appear shifted by 1. 

Fixed this issue by adding a dummy first parameter to the list of command
line arguments passed to d2. 

Hopefully this fixes the issue with ticket:2 cut in the system.

Trey: Please verify this fix and close ticket:2 in Trac.