Increased the memory checking threshold to 4.5 times the file size. I know it sounds ridiculous that it takes 4.5 times more memory to display text in Java but that is the price we pay for Java and graphics.
Hopefully that will catch any potential issues with JVM running out of 
memory and making it appear as if the GUI is hanging. This check-in fixes
the bug reported in ticket:12 in trac.

Added code to wrap file stream into a buffered stream to try and make the
reads go faster. Hopefully it will improve performance of loading large
files into the GUI.