Updated all the help buttons and URLs to directly point to various pages in the UserManual.pdf off peace-tools.org.
The solution more or less works fine in Windows and Linux. I have not tested
it under Mac. But typically, if it works on 2 of the most widely used operating
systems, it is safe to assume that it will *not* work on Mac.

However, I am not 100% satisfied about this solution due to the following
reasons:

1. It merely points to a page in the PDF. If the user manual is edited and
   pagination changes, then the URLs are going to be off.

2. I would have prefered to refer to individual HTML pages that are part of
   a larger HTML document. Unfortunately, the HTML generated by M$-Word is
   terrible and is practically unusable.

A long term solution would be to develop the user manual in LaTex or as a 
docbook (rather than as a Word document) so that it is convenient to export 
it in various formats.

This check-in will resolve ticket:6 in Trac. However, I will create a 
more long running echancement ticket to generate a better documentation
format that is more amenable to referencing from the GUI.