This check-in provides a significantly enhanced infrastructure for filtering out FASTA entries with low complexity segments. This infrastructure has been implemented in the form of Filters.
Filters are a new concept that have been added to PEACE. The
filters have been implemented in a similar manner as heuristics.
Accordingly, I have added the following classes:

Filter: The base class for all filters. It provides common
interface methods to interact with all filters.

FilterFactory: Centralized factory class for instantiating
various filters.

FilterChain: A chain of filters through which an ESTs must pass
in order to be included in the core clustering operation. This
class also does the work of applying filters and all-to-all
broadcasting of filtered ESTs to all processes participating in
an MPI process.

LengthFilter: A simple filter class that filters out ESTs that
are shorter than a given length. As a side effect ti fixes a bug
with earlier EST filtering logic.

Each filter class adds the ESTs it filters out to a dummy
cluster. The clusters now come out with names associated with
them (real clusters don't have a name) that indicate the reason
why the EST was placed in that cluster.

PEACE now provides option to export ESTs that pass and/or fail
filters into separate files for further processing or for feeding
into other tools. This feature will help us clean up FASTA
files. Needless to say, a --filter-only flag is included to have
PEACE just filter ESTs (in parallel of course) and do nothing
else.

I am checking-in the code so that hopefully it gets some
testing. Next I will add a filter that will weed out FASTA
entries with low complexity regions.