Added code to normalize base characters in a given nucleotide sequence read from a FASTA file using the following rules:
The EST::normalize() method normalizes the sequences such that the
resulting sequence is over the set {'A', 'T', 'C', 'G', 'N'}
in the following manner:

    - If the maskBases flag is true, then all low-case
      nucleotides are converted to 'N'. Otherwise they are converted
      to upper-case equivalents.<

    - All nucleotides that are not in "ATCG" are converted to 'N'

The normalization operation is done by default. It can be
suppressed by specifying --no-mask-bases. In this case (with
--no-mask-bases flag in command line arguments) "atcg" is
converted to "ATCG" but all other base characters (like 'X', 'n'
etc.) are converted to 'N'.

The idea is to ignore all the 'N' characters when running
heuristics, d2, and performing clustering.

Wonder how we would handle 'N' in Baton approach.