Updated the ESTCodec to now detect and flag 'N' characters (recollect from previous commit log PEACE now converts some base characters to 'N') in EST sequences. This is done using a third parameter in the operator() in functors NormalEncoder and RevCompEncoder.
I intentionally introduced the third parameter, thereby breaking
the code that uses them across the system. Now, the compiler
reports every spot that needs to be modified.  I methodically
modified these parts to check and ignore hash values that involve
an 'N' character.

This approach was proposed by Dr. Karro so that the system has a
neutral behavior for words with 'N' -- that is penalize nor favor
words with 'N' in them. Quoting Dr. Karro's example:

For the fragment:

   {{{ AAAAANCCCCCCNAAAAA }}}

(5 As, 1 N, 6 Cs, 1 N, 5 As)

For a word size of 6, only the word {{{ CCCCCC }}} would be put
into the hash.

Next, I resolved all the compile issues by suitably modifying the
core heuristics and distance computation algorithms (D2,
TwoPassD2, and D2Zim) to be compliant with the new API of
ESTCodec class. Most of the changes are relatively
straightforward except for two spots where I would like a second
pair of eyes to check. So I will be cutting an task ticket for
Trey in Trac to validate the two changes in heuristics.