Provided a slightly different implementation for D2. I used a std::vector<int> instead of int* for wordTables. The objective was to keep the core d2 implementation the same -- but just use (or add) only the words that were not tainted with a 'N' to the list of words to be processed.
My initial hope was to implement it without any if-statements and keep the core 
d2 loop the same (without checks for any sentinel values). This implementation
almost achieves that. 

There was only one place in which an if-statement was needed was when building 
the initial window (to handle two fragments that were shorter than the frame
size. It seemed appropriate to have an if-statement in the loop (rather than 
duplicatig a part of the for-loop) because this part is not the most 
heavily used part of the code.

Nevertheless, IMO, this approach seems to streamline the code a lot and got 
rid of explicit dynamic memory usage and made the D2 implementation look so 
much cleaner.