The initialization of parameters (via call to ParameterSetManager::setParameters()) for twoPassD2 needs to occur before initialization of twoPassD2 (itself). Maybe this is a check that needs to go into twoPassD2, that is it checks to see if the ParameterSetManager has been initialized, if not it generates a warning and picks up default values.
Moved a couple of lines of code within #ifdef further below in the code.
I think the #ifdef can be replaced with if-checks to streamline the example.

In addition, the parameter processing of this example is grown to the point
that it needs to be streamlined via the arg_parser in PEACE.

IMHO, This example needs help from a computer science person :-)

This  fixes ticket:28