A few compiler warnings and compatiblity with GCC 4.4 hash maps:
As of gcc 4.4 they recommend using the default hasher for std::string instead
of providing a custom one as done earlier. This check-in uses a configure
variable/check to selectively include a string hasher in PEACE depending on
the compiler version and supported platform-specific configurations.

This change needs to be verified on different platforms with different
compilers etc. It should not affect runtime functionality in any way.