Added an instance variable called sequenceLen along with an accessor method called getSequenceLength() to EST.  Now, the code uses the getSequenceLength() method [that merely returns the sequenceLen] instance variable rather than calculating the sequence length each time (by calling strlen).
The above change helped to save a sizable amount of CPU time as
string length was a frequently used operation in TVHeuristic's
inner-most loop.

With the above modification we are a tad bit faster than
wcd-0.5.1!  CAUTION: The above timings were obtained from a
synthetic FASTA file with 9000 randomly generated EST sequences
all about 1000 bp in length. See ticket:49 for additional
information about this data file.  Therefore, there wasn't much
adaptive analysis going on.  However, when running with long
reads (which is what I believe wcd is geared for) we should have
a similar situation (where there will not be a need for adaptive
analysis) and in those tests I hope we will be at least as good
as wcd.

The following numbers were obtained on Red Hawk running on a
single compute node:

wcd-0.5.1:
real 1m19.460s
user 0m0.034s
sys 0m0.284s

PEACE (SVN Revision: 652) -- i.e., latest
real	1m4.020s
user	0m0.098s
sys	0m0.884s

PEACE (0.9 earlier NAR version):
real	0m44.223s
user	0m0.099s
sys	0m0.778s


Notice that the earlier version of PEACE still runs much faster
than the current version. However, the primary reason for the
slow-down seems to be associated with handling 'X', 'N', and
other masked-bases where we compute hashes --- which occurs in
the core loop of PEACE as indicated below (in the earlier version
we did not have the if-check but just the body of the
if-statement shown below.  Similar code fragments occur in
several spots in the core analysis loops of PEACE causing some
slow down):

{{{
        if (wordIn != NHash) {
            score -= (delta[wordIn] << 1) - 1;
            delta[wordIn]--;
        }

}}}

The aforementioned if-check(s) seem to have added a good 40%
overhead in this test.  Maybe our next focus on reducing CPU
cycles needs to focus on this aspect of PEACE, if it is possible
to optimize.