Added a --dontAdapt flag to TwoPassD2.  By default TwoPassD2 will use adaptive parameters and will choose window (aka frame) size suitable to analyze a pair of reads based on their lengths -- that is, a pair of short reads get smaller windows while a pair of longer reads get larger window.  The adaptation enables optimal performance without compromising quality.
However, the performance does degrade when comparing PEACE (running slower due 
to small windows) with wcd (running with larger windows).  With the --dontAdapt
flag, the adaptation is turned off all the reads use a window size of 100,
the default we use.  This should provide comparable performance (hopefully) and
further tests are needed to verify that adaptation is really causing the
slowdown in all cases.

This check-in partially resolves ticket:52.  GUI changes are needed to 
expose the --dontAdapt flag (and after the GUI changes ticket:52 can be closed).