This check-in is an aggregate one that includes fixes for several tickets and bugs reported as a part of the NAR paper review. The specific issues addressed in this check-in are:
This check-in includes a Standard Flowgram Format (SFF) reader
that can read 454 data directly from SFF files.  The GUI uses the
SFF reader to load and process data.  Data sets in a workspace
now are explicitly tagged with the data file type to ease
processing. The workspace XML schema has been updated to include
the attribute. The modifications to XML schema are backwards
compatible and workspaces are automatically updated to the new
version when they are saved from the new version of PEACE GUI.
Note that currently, we only use the sequence information from
the SFF and don't use the per-base quality value.  However, this
information could become useful later on and we will need to
revisit the SFF reader again then. This check-in completely
resolves ticket:44.

I have updated the manual to indicate some basic configuration
details on setting up hosts file for MPICH and briefly mentioned
the use of `mpd` and `mpdboot` for running MPICH on a local
machine.  In have added links to MPICH and OpenMPI user manuals
for additional configuration information.  I have added a short
sentence in the GUI (at the appropriate location) referring the
user to the manual for configuration details.  These changes
resolve ticket:48.

I regenerated the Table Of Contents (TOC) for the user
manual (after changes), regenerated the PDF, and updated the
online manual. I am also checking-in the source MS-Word doc for
the user manual into the archive. This resolves ticket:46.

I have modified the server configuration page in the Job wizard
to report/accept memory on a per-process basis rather than as an
aggregate total. The GUI now computes the total memory (as:
Nodes * CPUsPerNode * MemPerCPU) before creating the job. This
resolves ticket:51.

I have updated the link to launch the job wizard in
`welcome.html`, the integrated quick start guide. The job wizard
correctly launches when the icon or the link is clicked. This
resolves ticket:41.

I added the new `AdaptiveTwoPassD2` class to the windows build as
well and I have rebuilt the windows executable. The updated
clustering engine tar balls have also been incorporated into the
GUI.

I have updated the version information at appropriate spots in
preparation for the upcoming release for second round of NAR
reviews.