Changes for ticket:60 :
I am done with incorporating the changes. For some reason I also had to 
move the earlier version of `MSTClusterMaker` class as well into the new 
code-path to achieve the earlier performance. I took a look at the differences 
but they seem minor; yet there was a notable performance difference that could
be bridged only with the earlier version of `MSTClusterMaker` that is now in 
the main branch under `OldMSTClusterMaker`. The command line to run all the 
earlier version of the code path is:

{{{

./peace --clusterMaker na-mst --analyzer oldTwopassD2 --heuristics oldTV --fastaFile ~/temp/pipeline/estsim_all_zf_cdnas.fa

}}}

It does not affect the new/current code path much at all. However, there is a 
minor modification that was needed to handle 'N's the same way as wcd does. 
Changes to earlier code was just formatting/indentation (as it appears that
programmers are using slightly different .emacs files with different tab-stop 
settings, mine is 4 while some others are using 8)

The next step would be to verify if the earlier code path provides the same 
performance on larger data sets (as way before) without significantly impacting
quality of Sanger reads (the earlier code path was meant only for Sanger reads
and not for short reads).