Here is the (much awaited?) check-in of  the parallel implementation for the Baton approach.  Now, PEACE finally has an integrated assembler component thereby meeting the promise of its title --- Parallel Environment for Assembly and Clustering of gene Expressions (PEACE).
The implementation has been integrated with PEACE and it utilizes some
of the infrastructure that was already present in PEACE.
Consequently, some of the defaults that PEACE uses are not completely
suitable to the baton approach. Some more code organization is
required to fully streamline assembly and clustering operations.
Meanwhile, the baton assembler can be run using the following command
line (As of now, there is no GUI support for running the baton
assembler):

mpiexec -n 4 ./peace --assembler baton --filters null --heuristics null --fastaFile test.fasta

The current baton approach implementation in PEACE must be viewed as
"preliminary" and it needs a lot of testing to validate that it is
operating correctly.  The implementation also permits the baton
approach to be used as an est analyzer to do pure clustering. Some
very basic testing indicates the code works fine in this mode as well.

There are a few changes to the infrastructure to be noted:

1. There is now an Assembler class hierarchy.  This hierarchy is
   similar to the earlier ESTAnalyzer hierarchy.

2. There is an AssemblerFactory to assist with creation of various
   assemblers, although the factory currently only creates one, the
   baton assembler.

3. I have refactored some of the methods used to load data into a
   FileLoader class so that the ESTAnalyzer and Assembler class
   hierarchies can share the file loading features.

4. Added some helper methods to ESTCodec to help with encoding and
   decoding operations in the assembler.

5. The baton assembler code path has been placed in its own
   sub-directory for convenience and improve code organization.  The
   code is heavily documented and programmer documentation in HTML
   (similar to what we currently have at
   http://www.peace-tools.org/prog_docs/c++_html/) can be generated
   using doxygen.  Some time down the line I will be generating and
   updating the HTML documentation online.

Note that this check-in also includes a few miscellaneous updates
(such as: formatting, indenting etc.) to various files. I guess we
need to share a common .emacs file for editing the code so that
we consistently use tab width of 4.

As always if you notice issues, need enhancements, or would like to
track tasks, please cut tickets via the Trac website at:
https://peace.csi.muohio.edu/projects/peace