Reverted back the "short-term" fix that Trey had in place for a bug (reported by Jens via personal communication) in Revision 714.
Instead, I made an equivalent small change to the length of the dummy 
sequences generated by the LC filter to not exceed the length of the longest 
sequence in the data set read.  This ensures that the D2 type algorithms 
and associated heuristics don't exceed the maximum buffer size (allocated 
in initialize() method) and the filter continutes to operate as before.