Finally, here comes a rather large and possibly sweeping set of check-ins to PEACE and GUI. It has been pending for a while -- I apologize for the delay and thank you all for your patience.  I am placing a brief overview of the changes further below. Please let me know if you have any questions or concerns. As always bug reports and feature requests need to be routed via Track at https://peace.csi.muohio.edu/trac/peace
PENDING Tasks
--------------
Here are some pending activities prior to releasing this version
of PEACE:

   * Ensure the C++ code compiles and runs under Mac OS-X
   * Update Windows build and ensure the C++ code builds and runs
     under Windows.
   * Testing on various platforms to ensure portability and
     interoperability with other supercomputing platforms.
   * Updating the user manuals
   * Updating peace-tools.org website


GUI Changes
-----------
The changes to the GUI is relatively incremental. The most
noticeable change is the integration of EAST into the GUI. The
GUI has some reasonable level of integration with EAST. The
functionality seems to work in a general case but needs further
testing to identify and iron out kinks.

Assembly & Clustering Engine Changes
------------------------------------

The brunt of the changes falls in reworking the clustering and
assembly infrastructure.  I have been working on re-factoring the
PEACE clustering & assembly parallel engine framework for a while
now. The parallel clustering & assembly (other than EAST) is now
organized as a collection of the following loosely coupled
subsystems: input subsystem, output subsystem, filtering
subsystem, clustering subsystem, and assembly subsystem. Files
for each of these subsystems live in their own directories. These
subsystems are controlled and coordinated by a core set of
classes constituting the framework (files in framework/
directory). Some of the shared data structures and utility
classes are in the shared/ directory. The top-level classes that
live directly in the src/ tree essentially trigger the various
operations in the framework.

There are many significant internal changes to the software
system. There are noticeable and hopefully provide a more
extensible framework as we move forward. However, the flexibility
and extensibility do come at some cost of increased software
complexity. However, the complexity merely reflects the
increasing complexity and sophistication of the software systems
we are attempting develop. So from that sense the increase in
complexity seems reasonable (at least to me).

From the user's functionality perspective the user should
experience a more streamlined command-line interface and
associated operations.  The framework provides mechanism to
minimize memory footprint (with some minimal performance
degradation) once this feature is fully coded (currently it is
not turned-on in the code). Using different input file formats
and output file formats should be relatively straightforward.