Added checks to ensure that data sets and MST files that were not valid are not presented as options to the user to avoid confusion.
In addition, a part of this check-in is related to fixing the order 
of EAST and PEACE jobs as well.

Tweak to the default/suggested output file names directory to be 
consistent with the job ID. In addition, changed the file-type extension
for the singletons file to be fasta rather than txt (as it is a fasta file).