There are a few changes to the clustering sub-system to fix a memory issue reported by Hazelhurst.
I added a class called NNMSTClusterMaker that attempts to use
nearest-neighbor (NN) to minimize number of heuristic and d2 
calls to improve performance of clustering. It does seem to 
improve clustering by 30% or so. However, it generates a slightly
different number of clusters. Need further testing and to see
how effective this is. Maybe we will get some student to puruse
this nearest-neighbor optimization at sometime.