Several big changes to the PEACE framework in preparation for primes-based heuristics.
The first big change is that MPI has been moved to use just the
C-bindings instead of C++ bindings.  This change was needed because
the C++ bindings have been deprecated and removed in OpenMPI (starting
with 3.0 version)

The next big change is that the code base has been moved to use C++11
as the minimum C++ requirement.  The bigest change is for HashMap,
which uses unordered_map.

Added an initial implementation for a Prime's based heuristics