Changes to the build system to streamline the libraries and compile process to enable integration with CUDA.
Earlier, the build system was using libtool, to build both static and
shared libraries.  However, the shared library feature was never
used. In addition, the shared library (with -fPIC) option was causing
compile issues with nvcc (which I realize now, nvcc is not very well
designed).

I have removed the use of libtool. Now the build system just
creates standard static libraries for various subsystems in PEACE.

Most of the changes are very straightforward to change "libpeace.la"
to "libpeace.a".  However, there are a copule of changes of
significance:

  1. configure.ac now includes checks to conditionally compile CUDA

  2. The Makefile.am in "clustering" directory has rules to build .cu files
  
  3. The libpeace.a is built using a custom rule to build the archive
     by combining files from the subsystem libraries.